Add files via upload
DAY 25 (QUESTION 19):
Your company received a lot of bad reviews about some of your products lately and the management
wants to see which products they are and how many have been returned so far. Using the Orders and returns table,
write a query to see the top 5 most returned products from the company.